Updated GLFW and glad, other compilation fixes

+++ src/Text2D.cpp
@@ -17,7 +17,7 @@ void Text2D::render_object() const {
     glDisable(GL_BLEND);
 }
 
-std::vector<Uv2DVertex> Text2D::gen_buffer(const glm::vec2 &position, int n, std::string_view text) {
+std::vector<Uv2DVertex> Text2D::gen_buffer(const glm::vec2 &position, float n, std::string_view text) {
     std::vector<Uv2DVertex> buffer(text.size());
     auto buffer_it{buffer.begin()};
